# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Keops Studios API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Game {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** e.g. Mobile, PC, Mobile & PC */
  platform: string;
  /** @nullable */
  genre?: string | null;
  /** e.g. In Development, Released, Coming Soon */
  status: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  storeUrl?: string | null;
  order: number;
  createdAt?: string;
}

export interface GameInput {
  title: string;
  description?: string;
  platform: string;
  genre?: string;
  status: string;
  imageUrl?: string;
  storeUrl?: string;
  order?: number;
}

export interface GameUpdate {
  title?: string;
  description?: string;
  platform?: string;
  genre?: string;
  status?: string;
  imageUrl?: string;
  storeUrl?: string;
  order?: number;
}

export interface TeamMember {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  roleTr?: string | null;
  /** @nullable */
  roleEn?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  bioTr?: string | null;
  /** @nullable */
  bioEn?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  order: number;
}

export interface TeamMemberInput {
  name: string;
  role: string;
  roleTr?: string;
  roleEn?: string;
  bio?: string;
  bioTr?: string;
  bioEn?: string;
  imageUrl?: string;
  linkedinUrl?: string;
  twitterUrl?: string;
  order?: number;
}

export interface TeamMemberUpdate {
  name?: string;
  role?: string;
  roleTr?: string;
  roleEn?: string;
  bio?: string;
  bioTr?: string;
  bioEn?: string;
  imageUrl?: string;
  linkedinUrl?: string;
  twitterUrl?: string;
  order?: number;
}

export interface SiteContent {
  id: number;
  /** @nullable */
  heroTitleTr?: string | null;
  /** @nullable */
  heroTitleEn?: string | null;
  /** @nullable */
  heroSubtitleTr?: string | null;
  /** @nullable */
  heroSubtitleEn?: string | null;
  /** @nullable */
  aboutTr?: string | null;
  /** @nullable */
  aboutEn?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  discordUrl?: string | null;
  /** @nullable */
  email?: string | null;
}

export interface SiteContentUpdate {
  heroTitleTr?: string;
  heroTitleEn?: string;
  heroSubtitleTr?: string;
  heroSubtitleEn?: string;
  aboutTr?: string;
  aboutEn?: string;
  instagramUrl?: string;
  twitterUrl?: string;
  linkedinUrl?: string;
  youtubeUrl?: string;
  discordUrl?: string;
  email?: string;
}

export interface AdminCredentials {
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
  /** @nullable */
  username?: string | null;
}

export interface StudioStats {
  gameCount: number;
  teamSize: number;
  yearsActive: number;
  platforms: string[];
}